requirements.txt
asgiref==3.3.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
Django==3.2
django-cors-headers==3.7.0
djangorestframework==3.12.4
drf-jwt==1.19.0
idna==2.10
pycparser==2.20
PyJWT==2.1.0
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.4
gunicorn==20.1.0
posts==0.2.2
psycopg2-binary==2.8.6
pytz==2021.1
whitenoise==5.2.0